requirements.txt
python-dotenv==1.0.1 # For reading environment variables
langchain
langchain-community
unstructured==0.14.4 # Document loading
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
onnxruntime
chromadb # Vector storage
tiktoken==0.7.0  # For embeddings 
pandas
matplotlib
seaborn
torch
transformers
sentence-transformers
pyyaml
psutil
hf_xet

# `pip install -U langchain-chroma`
# `pip install "unstructured[md]"`
